requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
resend